Fixing warnings

+++ TakeControl/takecontrol.Domain.UnitTests/Models/Players/PlayerXUnitTests.cs
@@ -16,7 +16,7 @@ public class PlayerXUnitTests
         var player = Player.Create(userId, name, 1, 1, 1);
 
         Assert.NotNull(player);
-        Assert.Equal((int)PlayerLevel.Begginer, (int)player.PlayerLevel);
+        Assert.Equal((int)PlayerLevel.Begginer, player.PlayerLevel);
         Assert.Equal(player.Name, name);
         Assert.Equal(player.UserId, userId);
     }
@@ -30,7 +30,7 @@ public class PlayerXUnitTests
         var player = Player.Create(userId, name, 1, 3, 2);
 
         Assert.NotNull(player);
-        Assert.Equal((int)PlayerLevel.Mid, (int)player.PlayerLevel);
+        Assert.Equal((int)PlayerLevel.Mid, player.PlayerLevel);
         Assert.Equal(player.Name, name);
         Assert.Equal(player.UserId, userId);
     }
@@ -44,7 +44,7 @@ public class PlayerXUnitTests
         var player = Player.Create(userId, name, 2, 4, 4);
 
         Assert.NotNull(player);
-        Assert.Equal((int)PlayerLevel.Expert, (int)player.PlayerLevel);
+        Assert.Equal((int)PlayerLevel.Expert, player.PlayerLevel);
         Assert.Equal(player.Name, name);
         Assert.Equal(player.UserId, userId);
     }
@@ -58,7 +58,7 @@ public class PlayerXUnitTests
         var player = Player.Create(userId, name, 2, 4, 4);
 
         Assert.NotNull(player);
-        Assert.Equal((int)PlayerLevel.Expert, (int)player.PlayerLevel);
+        Assert.Equal((int)PlayerLevel.Expert, player.PlayerLevel);
         Assert.Equal(player.Name, name);
         Assert.Equal(player.UserId, userId);
     }
@@ -71,7 +71,7 @@ public class PlayerXUnitTests
         var player = Player.Create(userId, null, 2, 4, 4);
 
         Assert.NotNull(player);
-        Assert.Equal((int)PlayerLevel.Expert, (int)player.PlayerLevel);
+        Assert.Equal((int)PlayerLevel.Expert, player.PlayerLevel);
         Assert.Null(player.Name);
         Assert.Equal(player.UserId, userId);
     }
